refactor: verify_token
